pyproject.toml
[tool.poetry]
name = "drf-confidential"
version = "0.1.1"
description = "Handle sensitive fields for django rest framework"
authors = ["Liquan Yang <yang.liquan87@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/resurrexi/drf-confidential"
repository = "https://github.com/resurrexi/drf-confidential"
classifiers = [
  "Framework :: Django",
  "Topic :: Software Development :: Libraries :: Python Modules"
]

[tool.poetry.dependencies]
python = "^3.7"
django = "<3.0"
djangorestframework = "^3.11.0"

[tool.poetry.dev-dependencies]
pytest = "^5.4.2"
pytest-django = "^3.9.0"
pytest-cov = "^2.8.1"
coverage = "^5.1"
codecov = "^2.0.22"
pylint = "^2.5.2"
pylint-django = "^2.0.15"
black = "^19.10b0"
six = "^1.15.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 79
target-version = ['py36', 'py37', 'py38']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | sunbeam/settings
  | __pycache__
  | migrations
)/
'''
